Add validate_transactions(), add copied_transactions, and fix bug
There was a bug with the way <amount_sent> and <amount_received> were calculated. <get_balance()> would return the correct balance for the first transaction, but not for ones after this. One side note: <transaction_sender> was defined like [block 1, block 2, etc] and [[transaction 1], [transaction 2], [etc]], but the original <amount_sent> was defined as if <transaction_sender> was [transaction 1, transaction 2, etc], making it less readable and confusing to debug.